# api.yaml
openapi: "3.0.3"
info:
  title: Anki Classroom API
  description: |
    Backend API for the Anki Classroom spaced repetition platform.
    All authenticated endpoints require a JWT in the `Authorization: Bearer <token>` header.
    Tokens are obtained via `POST /auth/login` and expire after 24 hours.
    Roles: `admin`, `teacher`, `student`.
  version: "0.1.0"
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SubmitReview:
      type: object
      required: [card_id, rating]
      properties:
        card_id: { type: integer }
        rating:
          {
            type: integer,
            minimum: 1,
            maximum: 4,
            description: "1=Again, 2=Hard, 3=Good, 4=Easy",
          }
        response_time_ms: { type: integer, nullable: true }
    ReviewResponse:
      type: object
      properties:
        card_id: { type: integer }
        state: { type: string }
        due_at: { type: integer, nullable: true }
        stability: { type: number, format: double }
        difficulty: { type: number, format: double }
        reps: { type: integer }
        lapses: { type: integer }
        interval_days: { type: integer, description: "Days until next review" }
...
    FlagResponse:
      type: object
      properties:
        card_id: { type: integer }
        flag: { type: integer }
    CardModResponse:
      type: object
      properties:
        card_id: { type: integer }
        suspended: { type: integer, description: "1 if suspended, 0 otherwise" }
        buried_until:
          { type: integer, nullable: true, description: "Unix seconds until card reappears; null if not buried" }
        due_at: { type: integer, nullable: true }
        state: { type: string }
    NoteModResponse:
      type: object
      properties:
        note_id: { type: integer }
        cards_affected: { type: integer, description: "Number of cards affected" }
        suspended: { type: integer, description: "1 if suspended, 0 otherwise" }
        buried_until:
          { type: integer, nullable: true, description: "Unix seconds until cards reappear; null if not buried" }
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /reviews:
    post:
      tags: [Reviews]
      summary: Submit a card review rating
      description: |
        FSRS recalculates the next review interval.
        Rating 1 (Again): due_at is set to now so the card re-appears
        in the study queue on the next fetch. The card cycles back until
        the student gives a passing rating (2-4).
        Ratings 2-4: the card gets a future due_at based on the FSRS interval.
      operationId: submitReview
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitReview"
      responses:
        "200":
          description: Review processed, new schedule returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResponse"
        "400":
          description: Rating must be 1-4, or card state not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          "401": { $ref: "#/components/responses/Unauthorized" }
          "404":
            description: Card state not found — start a study session first
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"

    /cards/{cardId}/flag:
      patch:
        tags: [Reviews]
        summary: Set or clear a card flag (Anki-style)
        description: "Flags are per-student markers. Values: 0=none, 1-7=colored flags."
        operationId: setFlag
        security: [{ bearerAuth: [] }]
        parameters:
          - { name: cardId, in: path, required: true, schema: { type: integer } }
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required: [flag]
                properties:
                  flag:
                    { type: integer, minimum: 0, maximum: 7, description: "Flag value 0-7" }
        responses:
          "200":
            description: Flag set
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/FlagResponse"
          "400": { $ref: "#/components/responses/BadRequest" }
          "401": { $ref: "#/components/responses/Unauthorized" }
          "404":
            description: Card state not found
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"

    /cards/{cardId}/suspend:
      post:
        tags: [Cards]
        summary: Suspend a card (exclude from study)
        operationId: suspendCard
        security: [{ bearerAuth: [] }]
        parameters:
          - { name: cardId, in: path, required: true, schema: { type: integer } }
        responses:
          "200":
            description: Card suspended
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/CardModResponse"
          "401": { $ref: "#/components/responses/Unauthorized" }
          "404":
            description: Card state not found
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"

    /cards/{cardId}/unsuspend:
      post:
        tags: [Cards]
        summary: Unsuspend a card
        operationId: unsuspendCard
        security: [{ bearerAuth: [] }]
        parameters:
          - { name: cardId, in: path, required: true, schema: { type: integer } }
        responses:
          "200":
            description: Card unsuspended
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/CardModResponse"
          "401": { $ref: "#/components/responses/Unauthorized" }
          "404":
            description: Card state not found
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"

    /cards/{cardId}/bury:
      post:
        tags: [Cards]
        summary: Bury a card until tomorrow
        operationId: buryCard
        security: [{ bearerAuth: [] }]
        parameters:
          - { name: cardId, in: path, required: true, schema: { type: integer } }
        responses:
          "200":
            description: Card buried until tomorrow
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/CardModResponse"
          "401": { $ref: "#/components/responses/Unauthorized" }
          "404":
            description: Card state not found
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"

    /cards/{cardId}/unbury:
      post:
        tags: [Cards]
        summary: Unbury a card immediately
        operationId: unburyCard
        security: [{ bearerAuth: [] }]
        parameters:
          - { name: cardId, in: path, required: true, schema: { type: integer } }
        responses:
          "200":
            description: Card unburied
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/CardModResponse"
          "401": { $ref: "#/components/responses/Unauthorized" }
          "404":
            description: Card state not found
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"

    /cards/{cardId}/reschedule:
      patch:
        tags: [Cards]
        summary: Manually set a card's due date
        description: "Provide either `due_at` (Unix seconds) or `days` (relative offset)."
        operationId: rescheduleCard
        security: [{ bearerAuth: [] }]
        parameters:
          - { name: cardId, in: path, required: true, schema: { type: integer } }
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  due_at:
                    { type: integer, description: "Absolute due timestamp (Unix seconds)" }
                  days:
                    { type: integer, description: "Relative due offset in days from now" }
        responses:
          "200":
            description: Card rescheduled
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/CardModResponse"
          "400": { $ref: "#/components/responses/BadRequest" }
          "401": { $ref: "#/components/responses/Unauthorized" }
          "404":
            description: Card state not found
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"

    /notes/{noteId}/suspend:
      post:
        tags: [Cards]
        summary: Suspend all cards belonging to a note
        description: "Bulk suspension of every card generated from a note, for the authenticated student."
        operationId: suspendNote
        security: [{ bearerAuth: [] }]
        parameters:
          - { name: noteId, in: path, required: true, schema: { type: integer } }
        responses:
          "200":
            description: Note cards suspended
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/NoteModResponse"
          "401": { $ref: "#/components/responses/Unauthorized" }
          "404":
            description: No cards found for this note
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"

    /notes/{noteId}/bury:
      post:
        tags: [Cards]
        summary: Bury all cards belonging to a note until tomorrow
        description: "Bulk burial of every card generated from a note, for the authenticated student."
        operationId: buryNote
        security: [{ bearerAuth: [] }]
        parameters:
          - { name: noteId, in: path, required: true, schema: { type: integer } }
        responses:
          "200":
            description: Note cards buried until tomorrow
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/NoteModResponse"
          "401": { $ref: "#/components/responses/Unauthorized" }
          "404":
            description: No cards found for this note
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"

  # -----------------------------------------------------------------
  # Deck Options (presets)
  # -----------------------------------------------------------------
...